+++ src/zukan_icon_theme/helpers/read_extract_zip.py
@@ -79,6 +79,4 @@ def extract_remove_folder(
 
     except Exception:
         shutil.rmtree(tempdir)
-        logger.exception(
-            'error occurred while extracting the sublime-package file.'
-        )
+        logger.exception('error occurred while extracting the sublime-package file.')